Created UniqueNameSet
