+++ Flame.Intermediate/Emit/UniqueNameMap.cs
@@ -9,50 +9,32 @@ namespace Flame.Intermediate.Emit
 {
     /// <summary>
     /// Defines a type that generates a unique name for every
-    /// value it is passed.
+    /// value it is passed. Generated names are stored, such
+    /// that requesting a name for the same object more than once
+    /// results in the same unique name.
     /// </summary>
     /// <typeparam name="T"></typeparam>
     public class UniqueNameMap<T>
     {
         public UniqueNameMap(Func<T, string> GetName, string Prefix)
-            : this(GetName, (elem, index) => Prefix + index)
-        { }
+        {
+            this.nameSet = new UniqueNameSet<T>(GetName, Prefix);
+            this.dict = new ConcurrentDictionary<T, string>();
+        }
         public UniqueNameMap(Func<T, string> GetName, Func<T, int, string> GenerateName)
         {
-            this.getName = GetName;
-            this.generateName = GenerateName;
+            this.nameSet = new UniqueNameSet<T>(GetName, GenerateName);
             this.dict = new ConcurrentDictionary<T, string>();
-            this.nameSet = new HashSet<string>();
         }
 
-        private Func<T, string> getName;
-        private Func<T, int, string> generateName;
+        private UniqueNameSet<T> nameSet;
         private ConcurrentDictionary<T, string> dict;
-        private HashSet<string> nameSet;
-
-        private string createName(T Element)
-        {
-            string name = getName(Element);
-
-            if (nameSet.Add(name))
-            {
-                return name;
-            }
-            int index = 0;
-            do
-            {
-                name = generateName(Element, index);
-                index++;
-            } while(!nameSet.Add(name));
-
-            return name;
-        }
 
         public string this[T Element]
         {
             get
             {
-                return dict.GetOrAdd(Element, createName);
+                return dict.GetOrAdd(Element, nameSet.GenerateName);
             }
         }
     }
